Lots of minor changes to prevent warnings. Small fixes to Path, AudioThumbnail, ValueTree, OutputStreams. Tweaks to convert Drawables to/from ValueTrees. New method to write XML to a stream.

+++ src/audio/audio_file_formats/juce_AudioCDReader.cpp
@@ -133,7 +133,7 @@ void AudioCDReader::refreshTrackLengths()
 
             if (r != 0)
             {
-                sample += r->lengthInSamples;
+                sample += (int) r->lengthInSamples;
                 delete r;
             }
         }
